基本实现sharding插件的事务代码及并发测试集

+++ internal/protocol/mysql/plugin/forward/handler.go
@@ -5,9 +5,10 @@ import (
 	"errors"
 	"fmt"
 
+	"github.com/ecodeclub/ekit/sqlx"
 	"github.com/ecodeclub/ekit/syncx"
 	"github.com/meoying/dbproxy/internal/datasource"
-	"github.com/meoying/dbproxy/internal/protocol/mysql/driver/sharding"
+	"github.com/meoying/dbproxy/internal/datasource/transaction"
 	"github.com/meoying/dbproxy/internal/protocol/mysql/internal/pcontext"
 	"github.com/meoying/dbproxy/internal/protocol/mysql/internal/visitor/vparser"
 	"github.com/meoying/dbproxy/internal/protocol/mysql/plugin"
@@ -30,12 +31,10 @@ func newHandler(ds datasource.DataSource) *Handler {
 
 func (h *Handler) Handle(ctx *pcontext.Context) (*plugin.Result, error) {
 	visitor := vparser.NewCheckVisitor()
-	stmtType := visitor.Visit(ctx.ParsedQuery.Root).(string)
-	switch stmtType {
-	case vparser.SelectStmt:
-		return h.handleSelectStmt(ctx)
-	case vparser.InsertStmt, vparser.UpdateStmt, vparser.DeleteStmt:
-		return h.handleCUDStmt(ctx)
+	sqlTypeName := visitor.Visit(ctx.ParsedQuery.Root).(string)
+	switch sqlTypeName {
+	case vparser.SelectStmt, vparser.InsertStmt, vparser.UpdateStmt, vparser.DeleteStmt:
+		return h.handleCRUDStmt(ctx, sqlTypeName)
 	case vparser.StartTransactionStmt:
 		return h.handleStartTransactionStmt(ctx)
 	case vparser.CommitStmt:
@@ -43,32 +42,42 @@ func (h *Handler) Handle(ctx *pcontext.Context) (*plugin.Result, error) {
 	case vparser.RollbackStmt:
 		return h.handleRollbackStmt(ctx)
 	default:
-		return nil, fmt.Errorf("%w", errors.New(stmtType))
+		return nil, fmt.Errorf("%w", errors.New(sqlTypeName))
 	}
 }
 
-// handleSelectStmt 处理Select语句
-func (h *Handler) handleSelectStmt(ctx *pcontext.Context) (*plugin.Result, error) {
-	var rows *sql.Rows
+// handleCRUDStmt 处理Select、Insert、Update、Delete操作
+func (h *Handler) handleCRUDStmt(ctx *pcontext.Context, sqlTypeName string) (*plugin.Result, error) {
+	var rows sqlx.Rows
+	var res sql.Result
 	var err error
-	if tx := h.getTxByConnID(ctx.ConnID); tx != nil {
-		rows, err = tx.Query(ctx, datasource.Query{
+	if sqlTypeName == vparser.SelectStmt {
+		rows, err = h.getDatasource(ctx).Query(ctx, datasource.Query{
 			SQL:  ctx.Query,
 			Args: ctx.Args,
 		})
 	} else {
-		rows, err = h.ds.Query(ctx, datasource.Query{
+		res, err = h.getDatasource(ctx).Exec(ctx, datasource.Query{
 			SQL:  ctx.Query,
 			Args: ctx.Args,
 		})
 	}
-
 	return &plugin.Result{
-		Rows: rows,
+		Rows:          rows,
+		Result:        res,
+		InTransaction: h.getConnTransactionState(ctx.ConnID),
 	}, err
 }
 
-// getTxByConnID 根据客户端连接ID获取事务, 因为事务是与链接绑定的
+// getDatasource 获取本次执行需要使用的数据源
+func (h *Handler) getDatasource(ctx *pcontext.Context) datasource.DataSource {
+	if tx := h.getTxByConnID(ctx.ConnID); tx != nil {
+		return transaction.NewTransactionDataSource(tx)
+	}
+	return h.ds
+}
+
+// getTxByConnID 根据客户端连接ID获取事务对象, 因为事务是与链接绑定的
 func (h *Handler) getTxByConnID(connID uint32) datasource.Tx {
 	if tx, ok := h.connID2Tx.Load(connID); ok {
 		return tx
@@ -76,36 +85,20 @@ func (h *Handler) getTxByConnID(connID uint32) datasource.Tx {
 	return nil
 }
 
-// handleCUDStmt 处理Insert、Update、Delete操作
-func (h *Handler) handleCUDStmt(ctx *pcontext.Context) (*plugin.Result, error) {
-	var err error
-	var res sql.Result
-	if tx := h.getTxByConnID(ctx.ConnID); tx != nil {
-		// 事务中
-		res, err = tx.Exec(ctx, datasource.Query{
-			SQL:  ctx.Query,
-			Args: ctx.Args,
-		})
-	} else {
-		res, err = h.ds.Exec(ctx, datasource.Query{
-			SQL:  ctx.Query,
-			Args: ctx.Args,
-		})
-	}
-
-	return &plugin.Result{
-		Result: res,
-	}, err
+// getConnTransactionState 根据客户端连接ID获取链接的事务状态
+func (h *Handler) getConnTransactionState(connID uint32) bool {
+	_, ok := h.connID2Tx.Load(connID)
+	return ok
 }
 
 // handleStartTransactionStmt 处理开启事务语句
 func (h *Handler) handleStartTransactionStmt(ctx *pcontext.Context) (*plugin.Result, error) {
-	tx, err := h.ds.BeginTx(sharding.NewSingleTxContext(ctx), nil)
+	tx, err := h.ds.BeginTx(transaction.UsingTxType(ctx, transaction.Single), nil)
 	if err != nil {
 		return nil, err
 	}
 	h.connID2Tx.Store(ctx.ConnID, tx)
-	return &plugin.Result{TxInTransaction: true}, nil
+	return &plugin.Result{InTransaction: true}, nil
 }
 
 // handleCommitStmt 处理提交事务语句
@@ -115,7 +108,9 @@ func (h *Handler) handleCommitStmt(ctx *pcontext.Context) (*plugin.Result, error
 	if tx != nil {
 		err = tx.Commit()
 	}
-	h.connID2Tx.Delete(ctx.ConnID)
+	if err == nil {
+		h.connID2Tx.Delete(ctx.ConnID)
+	}
 	return &plugin.Result{}, err
 }
 
@@ -126,6 +121,8 @@ func (h *Handler) handleRollbackStmt(ctx *pcontext.Context) (*plugin.Result, err
 	if tx != nil {
 		err = tx.Rollback()
 	}
-	h.connID2Tx.Delete(ctx.ConnID)
+	if err == nil {
+		h.connID2Tx.Delete(ctx.ConnID)
+	}
 	return &plugin.Result{}, err
 }
